+++ dxr/tests/trybuild/enum.rs
@@ -1,0 +1,8 @@
+use dxr::FromDXR;
+
+#[derive(FromDXR)]
+pub enum Enum {
+    Variant,
+}
+
+fn main() {}