+++ dxr/tests/trybuild/recursive.rs
@@ -1,0 +1,8 @@
+use dxr::{FromDXR, ToDXR};
+
+#[derive(FromDXR, ToDXR)]
+pub struct Recursive {
+    recursive: Box<Recursive>,
+}
+
+fn main() {}