+++ dxr/tests/trybuild/toref.rs
@@ -1,0 +1,8 @@
+use dxr::FromDXR;
+
+#[derive(FromDXR)]
+pub struct Reference<'a> {
+    string: &'a str,
+}
+
+fn main() {}