+++ dxr/tests/trybuild/tuple.rs
@@ -1,0 +1,6 @@
+use dxr::FromDXR;
+
+#[derive(FromDXR)]
+pub struct TupleStruct(String);
+
+fn main() {}